backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
langchain==0.0.350
openai==1.3.5
pymysql==1.1.0
cx_Oracle==8.3.0
python-jose[cryptography]==3.3.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
google-cloud-bigquery==3.17.1
google-auth==2.28.1 